Add ActogramJ
ActogramJ is a plugin for visualizing and analyzing actograms, i.e. activity diagrams. It was intended for actograms from fruit fly experiments performed in the Department of Neurobiology and Genetics in Wuerzburg, but can be used for any actograms. It went public back in 2010, and is still in use in various research groups.
https://bene51.github.io/ActogramJ
https://github.com/bene51/ActogramJ