List mirrors of updates sites right after the original site.
The order that update sites matter when multiple update sites provide
the same files.  This is how the Java-8 update site overwrites files
provided by the ImageJ and Fiji update sites.

At the moment, the ImageJ, Fiji, and Java-8 mirror site are listed
after the original.  If someone selects the ImageJ and Fiji mirrors
but not the Java-8 mirror, then the Java-8 overwrite is replaced by
the mirrors.  To avoid this, this commit moves the ImageJ and Fiji
update sites under the original update site.  See
https://forum.image.sc/t/errors-with-european-mirrors-for-fiji-and-imagej/103355