Make ImageJ and Fiji the first two sites again
This partially reverts 3bd33ccdde3e1b0268deb065afbe01e8584f16da.
Unfortunately, the ImageJ Updater hardcodes the assumption that the
first update site is https://sites.imagej.net/ and the second one is
https://sites.fiji.sc/. If that ever changes, old installations will
throw an IOException about the list of update sites being invalid.

For the relevant code, see:
https://github.com/imagej/imagej-updater/blob/imagej-updater-1.0.1/src/main/java/net/imagej/updater/util/AvailableSites.java#L131-L143

So to pacify the Updater, this commit sets the list to once again
conform to this assumption, at least for now. We still list the ImageJ
and Fiji EU mirrors before the Java-8 update site though, in the hopes
that this results in fewer problems for folks than with Java-8 third.

Hopefully fixes fiji/fiji#358.